+++ src/transformers/models/minimax/modeling_minimax.py
@@ -460,7 +460,7 @@ class MiniMaxSparseMoeBlock(nn.Module):
         self.experts = MiniMaxExperts(config)
 
     def route_tokens_to_experts(self, router_logits):
-        routing_weights = torch.nn.functional.softmax(router_logits, dim=-1)
+        routing_weights = torch.nn.functional.softmax(router_logits.float(), dim=-1)
         top_k_weights, top_k_index = torch.topk(routing_weights, self.top_k, dim=-1)
         top_k_weights /= top_k_weights.sum(dim=-1, keepdim=True)
         return top_k_index, top_k_weights
