+++ src/transformers/models/mixtral/modeling_mixtral.py
@@ -115,7 +115,7 @@ class MixtralSparseMoeBlock(nn.Module):
         self.experts = MixtralExperts(config)
 
     def route_tokens_to_experts(self, router_logits):
-        routing_weights = torch.nn.functional.softmax(router_logits, dim=-1)
+        routing_weights = torch.nn.functional.softmax(router_logits.float(), dim=-1)
         top_k_weights, top_k_index = torch.topk(routing_weights, self.top_k, dim=-1)
         top_k_weights /= top_k_weights.sum(dim=-1, keepdim=True)
         return top_k_index, top_k_weights
